Move generate chain to tests.
